Time: 5 ms (33.70%), Space: 6.6 MB (91.95%) - LeetHub

+++ 377-combination-sum-iv/377-combination-sum-iv.cpp
@@ -36,7 +36,7 @@ public:
 
     int combinationSum4(vector<int>& nums, int target) 
     {
-        vector<unsigned long int> dp(target + 1);
+        vector<unsigned int> dp(target + 1);
         dp[0] = 1;
         sort (nums.begin(), nums.end());
         for (int i = 1; i <= target; i++)
